requirements.txt
APScheduler==3.10.4
blessed==1.20.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
docker==7.0.0
Flask==3.0.1
Flask-APScheduler==1.13.1
Flask-HTTPAuth==4.8.0
gpustat==1.1.1
GPUtil==1.4.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
Naked==0.1.32
nvidia-ml-py==12.535.133
packaging==23.2
psutil==5.9.8
pycparser==2.21
pycryptodome==3.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
zipp==3.17.0